Added functionality, bug fixes
Formatter.py

Changed way color is used for command prompt using color_console.py.
Neater chunk_trace() which displays the layer of the chunk to the left
of the trace. Fixed INDEX in get_dynamic_number() and added P in PEMDAS
for parentheses in get_dynamic_number(). Added new try except around
getting the num_chunks in the CHUNK if block. Added traceback for when
"ERROR" is assigned to raw values in get_raw().