Optimized and fixed format
bin_decode3.py

Is now safer in terms of handling errenous data.

config.txt was added back in as a way to interface with the program once
it has been frozen as an executeable.

formats/wdb

Got rid of coordinate_things, replaced with something more sensible.

formatter.py

Added a bunch of optimizations to cut extraction time down by half,
getting rid of trace statements and such with a setting at the top of
the file.

wdb_ripper.py

Got rid of pygame for handling pngs, switched to pypng. Added logging
options. Consolidated imports. Added support for the extra hidden images
in isle and isle_hi.

And got rid of travis.